refactor: Pulls out binary helpers into modules
